Made disabled login button have clearer text
Having a disabled login button with almost transparent text makes it hard to read, thus not as aesthetic as a clearer label. IMHO transparent text is not that good an indicator that the button is unclickable. Besides, everyone knows that they have to enter a password before they log in, as is evident by the password input, and by the fact that they created a password.